Moved `GetStatistics` from `usecase` package
